Added MacOS versions of FreeImage.NET
Added a universal Assembly which uses non unicode api
calls when loading files. This should work on all platforms
as the unicode versions only worked on windows.

The libfreeimage.dylib files were sourced from brew and
are universal binaries and as such will work on both
x64 and x86 system.